angularwithtests.client/src/app/customer-list/customer-list.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { CustomerListComponent } from './customer-list.component';
import { Customer } from '../Customer';
import { DataService } from '../data.service';

class MockDataService {
  customers: Customer[] = [{
    id: 0,
    firstName: 'Bob',
    lastName: 'Jones',
    phone: '5551212'
  }];

  addCustomer(customer: Customer) {
    this.customers.push(customer);
  }
}

describe('CustomerListComponent', () => {
  let component: CustomerListComponent;
  let fixture: ComponentFixture<CustomerListComponent>;
  let customer: Customer;
  let dataService: MockDataService;

  //arrange

  beforeEach(() => {
    dataService = new MockDataService();
    TestBed.configureTestingModule({
      declarations: [CustomerListComponent],
      providers: [{ provide: DataService, useValue: dataService }]
    });
    fixture = TestBed.createComponent(CustomerListComponent);
    component = fixture.componentInstance;

    customer = dataService.customers[0];

    fixture.detectChanges();
  });

  it('should create', () => {
    //act here

    //assert here
    expect(component).toBeTruthy();
  });

  it('should have a customer list', () => {
    component.customers.push(customer);

    expect(component.customers).toBeDefined();
    expect(component.customers).toEqual([customer]);
    expect(component.customers[0].firstName).toEqual("Bob");
  });

  it('should add a new customer when addCustomer is fired', () => {
    component.addCustomer();

    expect(component.customers.length).toEqual(3);
  });

  it('should have Customer List as the heading', () => {
    // find a specific element in our DOM

    const heading: HTMLElement = fixture.nativeElement;
    const h1 = heading.querySelector('h1')!;

    expect(h1.textContent).toEqual("Customer List");
    // expect that element to contain 'Customer List'
  });

});
